+++ packages/web/src/index.ts
@@ -9,8 +9,7 @@ export * from './registerStyles'
 // CodeZero UI components that extend Vaadin components
 export * from './animations/'
 export * from './banners/'
-export * from './buttons/button'
-export * from './buttons/saveButton'
+export * from './buttons/'
 export * from './details'
 export * from './dialogs/baseDialog'
 export * from './dialogs/baseStoreDialog'
